CASSANDRA-16524 Moved buffer allocation tests to the o.a.c.net package; Added overloaded method for `putUnusedPortion` that takes a `BufferPoolAllocator.Wrapped` object; Updated callers of `Wrapped.adopt()` to put unused bytes before adopting.
