+++ src/java/org/apache/cassandra/net/FrameDecoderLegacyLZ4.java
@@ -47,8 +47,6 @@ import static org.apache.cassandra.utils.ByteBufferUtil.copyBytes;
  */
 class FrameDecoderLegacyLZ4 extends FrameDecoderLegacy
 {
-    private static final BufferPool bufferPool = BufferPools.forNetworking();
-
     FrameDecoderLegacyLZ4(BufferPoolAllocator allocator, int messagingVersion)
     {
         super(allocator, messagingVersion);
@@ -123,9 +121,10 @@ class FrameDecoderLegacyLZ4 extends FrameDecoderLegacy
         public void channelRead(ChannelHandlerContext ctx, Object msg) throws CorruptLZ4Frame
         {
             assert msg instanceof BufferPoolAllocator.Wrapped;
-            ByteBuffer buf = ((BufferPoolAllocator.Wrapped) msg).adopt();
+            BufferPoolAllocator.Wrapped wrapped = (BufferPoolAllocator.Wrapped) msg;
             // netty will probably have mis-predicted the space needed
-            bufferPool.putUnusedPortion(buf);
+            allocator.putUnusedPortion(wrapped);
+            ByteBuffer buf = wrapped.adopt();
 
             CorruptLZ4Frame error = null;
             try
@@ -255,7 +254,7 @@ class FrameDecoderLegacyLZ4 extends FrameDecoderLegacy
             }
             catch (Throwable t)
             {
-                bufferPool.put(out);
+                allocator.put(out);
                 throw t;
             }
         }
@@ -272,7 +271,7 @@ class FrameDecoderLegacyLZ4 extends FrameDecoderLegacy
         {
             if (null != stash)
             {
-                bufferPool.put(stash);
+                allocator.put(stash);
                 stash = null;
             }
 
@@ -351,7 +350,7 @@ class FrameDecoderLegacyLZ4 extends FrameDecoderLegacy
             ByteBuffer out = allocator.getAtLeast(capacity);
             in.flip();
             out.put(in);
-            bufferPool.put(in);
+            allocator.put(in);
             return out;
         }
